gesture

+++ app/src/main/java/com/insightsurfface/videocrawler/listener/ProgressChangeListener.java
@@ -1,7 +1,7 @@
 package com.insightsurfface.videocrawler.listener;
 
 public interface ProgressChangeListener {
-    void onProgressChanged(int value);
+    void onProgressChanged(float value);
     void onStartTrackingTouch();
     void onStopTrackingTouch();
 }
